Adding error handling for different kind of requests

+++ src/main/java/hu/wup/restapitestproject/services/BookServiceImpl.java
@@ -1,6 +1,6 @@
 package hu.wup.restapitestproject.services;
 
-import hu.wup.restapitestproject.exceptions.BookNotFoundException;
+import hu.wup.restapitestproject.exceptions.ApiRequestException;
 import hu.wup.restapitestproject.model.Book;
 import hu.wup.restapitestproject.repositories.BookRepository;
 import org.apache.logging.log4j.LogManager;
@@ -28,11 +28,12 @@ public class BookServiceImpl implements BookService {
     @Override
     public Book getBookById(long id)  {
         Optional<Book> result = bookRepository.findById(id);
-        Book book = null;
+        Book book;
         if(result.isPresent()) {
             book = result.get();
         }else {
             logger.error("Book hasn't been found by this id:" + id);
+            throw new ApiRequestException(" Book has not been found by this id " + id);
         }
         logger.debug(result.get());
         return book;
@@ -42,31 +43,29 @@ public class BookServiceImpl implements BookService {
     @Override
     public void insertBook(Book book) {
         bookRepository.save(book);
-
     }
 
     @Override
     public void deleteBook(long bookId) {
-//        boolean exists = bookRepository.existsById(bookId);
-//        logger.debug("deleteBook by: " + bookId);
-//        if (!exists) {
-//            throw new IllegalStateException("Book with this id " + bookId + "does not exist.");
-//        }
+        boolean exists = bookRepository.existsById(bookId);
+        logger.debug("deleteBook by: " + bookId);
+        if (!exists) {
+            throw new ApiRequestException("Book does not exist by this id " + bookId);
+        }
         bookRepository.deleteById(bookId);
     }
 
     @Override
     public void updateBook(long bookId, Book book) {
 
-//        Book existingBook = bookRepository.findById(bookId).orElseThrow(
-//                () -> new IllegalStateException("Book with this id " + bookId + " does not exist!"));
-//        existingBook.setTitle(book.getTitle());
-//        existingBook.setAuthor(book.getAuthor());
-        bookRepository.save(book);
+        Book existingBook = bookRepository.findById(bookId).orElse(null);
+        logger.debug("updateBook by: " + bookId);
+        if (existingBook == null) {
+            throw new ApiRequestException("Book does not exist by this id " + bookId);
+        }else {
+         existingBook.setAuthor(book.getAuthor());
+         existingBook.setTitle(book.getTitle());
+        }
+        bookRepository.save(existingBook);
     }
-
-
-
-
-
 }
